Stable
I'll leave this here in this branch in case we ever need to go back to it, as a version we can leave untouched.

As far as I can tell, all the main things run correctly, and I tried to neaten up the code in main.py so it's easily understandable.